Add create Telegram notificator modal box in Gamecraft UI

+++ gamecraft-ui/src/main/resources/static/js/ui_operations.js
@@ -250,14 +250,26 @@ function fillTelegramBotsTable() {
     var bots = getTelegramBots();
     $(function() {
         $.each(bots, function(i, item) {
-            var tr = $('<tr>').append(
-                $('<tr>'),
-                $('<td>').text(item.id),
-                $('<td>').text(item.telegramBotName),
-                $('<td>').text(item.telegramBotDescription),
-                $('<td>').html("<button type=\"button\" class=\"btn btn-danger btn-xs\" onclick=\"deleteTelegramBot('" +item.id + "'); location.reload(); \"><span class=\"glyphicon glyphicon-remove\"></span> </button>")
-            );
-            $(".table").append(tr.html());
+            if (isAdmin(getUsername())) {
+                var tr = $('<tr>').append(
+                    $('<tr>'),
+                    $('<td>').text(item.id),
+                    $('<td>').text(item.telegramBotName),
+                    $('<td>').text(item.telegramBotDescription),
+                    $('<td>').html("<button type=\"button\" class=\"btn btn-danger btn-xs\" onclick=\"deleteTelegramBot('" + item.id + "'); location.reload(); \"><span class=\"glyphicon glyphicon-remove\"></span> </button>")
+                );
+                $(".table").append(tr.html());
+            }
+            else {
+                var tr = $('<tr>').append(
+                    $('<tr>'),
+                    $('<td>').text(item.id),
+                    $('<td>').text(item.telegramBotName),
+                    $('<td>').text(item.telegramBotDescription),
+                    $('<td>').html("")
+                );
+                $(".table").append(tr.html());
+            }
         });
     });
 }
